# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "FullStackDelas - Backend",
    "description": "API Backend do projeto FullStackDelas"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/linkedin": {
      "post": {
        "tags": [
          "Linkedin"
        ],
        "description": "Endpoint para cadastrar um registro.",
        "responses": {
          "200": {
            "description": "Registro cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseCreateSuccess"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/responseCreateSuccess"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/responseCreateSuccess"
            }
          },
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "example": "any"
                  },
                  "schema": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Linkedin"
        ],
        "description": "Endpoint para retornar todos registro.",
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/linkedin/{id}": {
      "get": {
        "tags": [
          "Linkedin"
        ],
        "description": "Endpoint para obter um registro por ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Linkedin"
        ],
        "description": "Endpoint para atualizar um registro por ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "linkedin": {
                    "example": "any"
                  },
                  "frase": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Linkedin"
        ],
        "description": "Endpoint para excluir um registro por ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "requestBody": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Nome do Usuário"
          },
          "email": {
            "type": "string",
            "example": "email@test.com"
          },
          "linkedin": {
            "type": "string",
            "example": "https://www.linkedin.com/in/usuario"
          },
          "frase": {
            "type": "string",
            "example": "Uma frase bonita"
          }
        },
        "required": [
          "nome",
          "email",
          "linkedin",
          "frase"
        ],
        "xml": {
          "name": "requestBody"
        }
      },
      "responseCreateSuccess": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "example": "Registro criado com sucesso"
          }
        },
        "required": [
          "msg"
        ],
        "xml": {
          "name": "responseCreateSuccess"
        }
      }
    }
  }
}